Fix #1c Rename the shared folder
*Renamed the Shared folder to shared-temp
*Moved the shared-temp folder into the app folder
