+++ src/app/header/header.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { LoginService } from '../../shared/login.service';
+import { LoginService } from '../shared_temp/login.service';
 import { User } from '../users/user.model';
-import { UserService } from '../../shared/user.service';
+import { UserService } from '../shared_temp/user.service';
 
 @Component({
   selector: 'app-header',
