+++ src/app/login/login.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { NgForm } from '@angular/forms';
-import { LoginService } from '../../shared/login.service';
+import { LoginService } from '../shared_temp/login.service';
 import { User } from '../users/user.model';
 import { Router } from '@angular/router';
-import { UserService } from '../../shared/user.service';
+import { UserService } from '../shared_temp/user.service';
 import { AlertService } from '../_alert';
