+++ src/app/projects/project-details/project-edit/project-edit.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { Project } from '../../project.model';
 import { NgForm } from '@angular/forms';
-import { ProjectService } from '../../../../shared/project.service';
+import { ProjectService } from '../../../shared_temp/project.service';
 import { ActivatedRoute, Router, Params } from '@angular/router';
 
 @Component({
