+++ src/app/projects/project-new/project-new.component.ts
@@ -1,8 +1,6 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { ActivatedRoute, Router, Params } from '@angular/router';
 import { NgForm } from '@angular/forms';
-
-import { ProjectService } from '../../../shared/project.service';
+import { ProjectService } from '../../shared_temp/project.service';
 
 
 @Component({
