+++ src/app/shared_temp/task.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Subject } from 'rxjs';
-import { Task } from '../app/tasks/task.model';
-import { User } from '../app/users/user.model';
-import { Project } from '../app/projects/project.model';
+import { Task } from '../tasks/task.model';
+import { User } from '../users/user.model';
+import { Project } from '../projects/project.model';
 import { ProjectService } from './project.service';
 import { UserService } from './user.service';
 import * as _ from "lodash";
