+++ src/app/shared_temp/user.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Subject } from 'rxjs';
-import { User } from '../app/users/user.model';
+import { User } from '../users/user.model';
 import { v4 as uuidv4 } from 'uuid';
 import * as _ from "lodash";
 import {Md5} from 'ts-md5/dist/md5';
