+++ src/app/tasks/task-new/task-new.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { NgForm } from '@angular/forms';
-import { TaskService } from '../../../shared/task.service';
+import { TaskService } from '../../shared_temp/task.service';
 
 @Component({
   selector: 'app-task-new',
