+++ src/app/users/user-new/user-new.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { UserService } from '../../../shared/user.service';
+import { UserService } from '../../shared_temp/user.service';
 import { ActivatedRoute, Router, Params } from '@angular/router';
 import { NgForm } from '@angular/forms';
